fix(usage): better audit/boolean flag usage output
This adds the `audit` config item to the usage output of `npm ci`, and
also tweaks how usage flags are shown for boolean options that do not
default to false.  Their usage is shown as the `--no-x` form of the flag
to better communicate that the flag is needed to turn that normally true
option OFF.

The description of `audit` was also updated to reflect that it runs on
many different npm commands, not just install.  Because this flag is
included in the usage of those commands it's best to let the assocation
happen there instead of trying to be comprehensive in the description
itself.

A small fix to make `install-ci-test` not try to redefine its own
usage, but to inherit from `ci` was also included.